LPS-57199 Move classes
