+++ modules/apps/staging/staging-taglib/src/META-INF/resources/taglib/portlet_list/render_controls.jsp
