+++ modules/apps/staging/staging-taglib/src/com/liferay/staging/taglib/MenuTag.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.taglib.staging;
+package com.liferay.staging.taglib;
 
 import com.liferay.taglib.util.IncludeTag;
